symbolic link needs the user check

+++ install.sh
@@ -5,15 +5,22 @@ case "${OSTYPE}" in
     brew install git
     ;;
   linux*)
-    sudo apt-get install software-properties-common
-    sudo apt-get update
-    sudo apt-get install neovim
+    sudo apt update
+    sudo apt install software-properties-common
+    sudo apt install neovim
     ;;
 esac
 
+read -p "Activate symbolic link of .vimrc? (y/n) :" YN
+if [ "${YN}" = "y" ]; then 
+    ln -s ~/dotfiles/.vimrc ~/.vimrc
+fi
 
-ln -s ~/dotfiles/.vimrc ~/.vimrc
-ln -s ~/dotfiles/.bashrc ~/.bashrc
+read -p "Activate symbolic link of .bashrc? (y/n) :" YN
+if [ "${YN}" = "y" ]; then 
+    ln -s ~/dotfiles/.bashrc ~/.bashrc
+fi
+
 ln -s ~/dotfiles/.git-completion.bash ~/.git-completion.bash
 ln -s ~/dotfiles/.git-prompt.sh ~/.git-prompt.sh
 
@@ -21,8 +28,12 @@ mkdir -p ~/.cache/dein/
 ln -s ~/dotfiles/.dein.toml ~/.cache/dein/.dein.toml
 ln -s ~/dotfiles/.dein_lazy.toml ~/.cache/dein/.dein_lazy.toml
 
-mkdir -p ~/.config/nvim
-ln -s ~/dotfiles/init.vim ~/.config/nvim/init.vim
+read -p "Activate neovim? (y/n) :" YN
+if [ "${YN}" = "y" ]; then 
+    ln -s ~/dotfiles/.bashrc ~/.bashrc
+    mkdir -p ~/.config/nvim
+    ln -s ~/dotfiles/init.vim ~/.config/nvim/init.vim
+fi
 
 mkdir -p ~/.vim/colors
 git clone https://github.com/tomasr/molokai
